+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.0.4-rc2] - 2026-05-18
+## [Unreleased]
 
 Upcoming maintenance release.
 
@@ -22,6 +22,8 @@ Upcoming maintenance release.
  - #30: Added CMake build configuration and CI workflows.
  
 ### Changed
+
+ - #31: Portability to Windows / MSC.
 
  - `examples/Makefile` to work standalone, without `config.mk`.